Move print implementation to source file
